+++ ppdet/modeling/assigners/task_aligned_assigner.py
@@ -163,7 +163,7 @@ class TaskAlignedAssigner(nn.Layer):
         assigned_gt_index = mask_positive.argmax(axis=-2)
 
         # assigned target
-        assigned_gt_index = assigned_gt_index + batch_ind * num_max_boxes
+        assigned_gt_index = assigned_gt_index + (batch_ind * num_max_boxes).astype(assigned_gt_index.dtype)
         assigned_labels = paddle.gather(
             gt_labels.flatten(), assigned_gt_index.flatten(), axis=0)
         assigned_labels = assigned_labels.reshape([batch_size, num_anchors])
